Added edm::core::OutputModuleCore
Consolidated all common code used by OutputModuleBase classes into one class.
